+++ src/main/kotlin/dev/aaronhowser/mods/excessive_utilities/block/EnderPorcupineBlock.kt
@@ -1,6 +1,6 @@
 package dev.aaronhowser.mods.excessive_utilities.block
 
-import dev.aaronhowser.mods.excessive_utilities.block_entity.EnderPorcupineBlockEntity
+import dev.aaronhowser.mods.excessive_utilities.block_entity.ender_porcupine.EnderPorcupineBlockEntity
 import net.minecraft.core.BlockPos
 import net.minecraft.world.InteractionResult
 import net.minecraft.world.MenuProvider